+++ spec/wavefront-cli/derivedmetric_spec.rb
@@ -79,3 +79,13 @@ describe "#{word} command" do
                 headers: JSON_POST_HEADERS }, k)
   test_list_output(word, k)
 end
+
+CliClass = WavefrontCli::DerivedMetric
+
+class TestDerivedMetricMethods < CliMethodTest
+  def test_import_method
+    import_tester(:derivedmetric,
+                  %i[tags minutes name query metricsUsed hostsUsed],
+                  %i[id])
+  end
+end